rumi_proto::google::protobuf

Struct FeatureSet

pub struct FeatureSet {
    pub field_presence: Option<i32>,
    pub enum_type: Option<i32>,
    pub repeated_field_encoding: Option<i32>,
    pub utf8_validation: Option<i32>,
    pub message_encoding: Option<i32>,
    pub json_format: Option<i32>,
}

Fields
field_presence: Option<i32>enum_type: Option<i32>repeated_field_encoding: Option<i32>utf8_validation: Option<i32>message_encoding: Option<i32>json_format: Option<i32>
Implementations

impl FeatureSet

pub fn field_presence(&self) -> FieldPresence

Returns the enum value of field_presence, or the default if the field is unset or set to an invalid enum value.

pub fn set_field_presence(&mut self, value: FieldPresence)

Sets field_presence to the provided enum value.

pub fn enum_type(&self) -> EnumType

Returns the enum value of enum_type, or the default if the field is unset or set to an invalid enum value.

pub fn set_enum_type(&mut self, value: EnumType)

Sets enum_type to the provided enum value.

pub fn repeated_field_encoding(&self) -> RepeatedFieldEncoding

Returns the enum value of repeated_field_encoding, or the default if the field is unset or set to an invalid enum value.

pub fn set_repeated_field_encoding(&mut self, value: RepeatedFieldEncoding)

Sets repeated_field_encoding to the provided enum value.

pub fn utf8_validation(&self) -> Utf8Validation

Returns the enum value of utf8_validation, or the default if the field is unset or set to an invalid enum value.

pub fn set_utf8_validation(&mut self, value: Utf8Validation)

Sets utf8_validation to the provided enum value.

pub fn message_encoding(&self) -> MessageEncoding

Returns the enum value of message_encoding, or the default if the field is unset or set to an invalid enum value.

pub fn set_message_encoding(&mut self, value: MessageEncoding)

Sets message_encoding to the provided enum value.

pub fn json_format(&self) -> JsonFormat

Returns the enum value of json_format, or the default if the field is unset or set to an invalid enum value.

pub fn set_json_format(&mut self, value: JsonFormat)

Sets json_format to the provided enum value.