rumi_proto::google::protobuf

Struct FieldOptions

pub struct FieldOptions {

    pub ctype: Option<i32>,
    pub packed: Option<bool>,
    pub jstype: Option<i32>,
    pub lazy: Option<bool>,
    pub unverified_lazy: Option<bool>,
    pub deprecated: Option<bool>,
    pub weak: Option<bool>,
    pub debug_redact: Option<bool>,
    pub retention: Option<i32>,
    pub targets: Vec<i32>,
    pub edition_defaults: Vec<EditionDefault>,
    pub features: Option<FeatureSet>,
    pub feature_support: Option<FeatureSupport>,
    pub uninterpreted_option: Vec<UninterpretedOption>,

}

Fields
ctype: Option<i32>packed: Option<bool>jstype: Option<i32>lazy: Option<bool>unverified_lazy: Option<bool>deprecated: Option<bool>weak: Option<bool>debug_redact: Option<bool>retention: Option<i32>targets: Vec<i32>edition_defaults: Vec<EditionDefault>features: Option<FeatureSet>feature_support: Option<FeatureSupport>uninterpreted_option: Vec<UninterpretedOption>
Implementations

impl FieldOptions

pub fn ctype(&self) -> CType

Returns the enum value of ctype, or the default if the field is unset or set to an invalid enum value.

pub fn set_ctype(&mut self, value: CType)

Sets ctype to the provided enum value.

pub fn packed(&self) -> bool

Returns the value of packed, or the default value if packed is unset.

pub fn deprecated(&self) -> bool

Returns the value of deprecated, or the default value if deprecated is unset.

pub fn lazy(&self) -> bool

Returns the value of lazy, or the default value if lazy is unset.

pub fn jstype(&self) -> JsType

Returns the enum value of jstype, or the default if the field is unset or set to an invalid enum value.

pub fn set_jstype(&mut self, value: JsType)

Sets jstype to the provided enum value.

pub fn weak(&self) -> bool

Returns the value of weak, or the default value if weak is unset.

pub fn unverified_lazy(&self) -> bool

Returns the value of unverified_lazy, or the default value if unverified_lazy is unset.

pub fn debug_redact(&self) -> bool

Returns the value of debug_redact, or the default value if debug_redact is unset.

pub fn retention(&self) -> OptionRetention

Returns the enum value of retention, or the default if the field is unset or set to an invalid enum value.

pub fn set_retention(&mut self, value: OptionRetention)

Sets retention to the provided enum value.

pub fn targets( &self, ) -> FilterMap<Cloned<Iter<'_, i32>>, fn(i32) -> Option<OptionTargetType>>

Returns an iterator which yields the valid enum values contained in targets.

pub fn push_targets(&mut self, value: OptionTargetType)

Appends the provided enum value to targets.