rumi_proto::google::protobuf

Struct FileOptions

pub struct FileOptions {

    pub java_package: Option<String>,
    pub java_outer_classname: Option<String>,
    pub java_multiple_files: Option<bool>,
    pub java_generate_equals_and_hash: Option<bool>,
    pub java_string_check_utf8: Option<bool>,
    pub optimize_for: Option<i32>,
    pub go_package: Option<String>,
    pub cc_generic_services: Option<bool>,
    pub java_generic_services: Option<bool>,
    pub py_generic_services: Option<bool>,
    pub deprecated: Option<bool>,
    pub cc_enable_arenas: Option<bool>,
    pub objc_class_prefix: Option<String>,
    pub csharp_namespace: Option<String>,
    pub swift_prefix: Option<String>,
    pub php_class_prefix: Option<String>,
    pub php_namespace: Option<String>,
    pub php_metadata_namespace: Option<String>,
    pub ruby_package: Option<String>,
    pub features: Option<FeatureSet>,
    pub uninterpreted_option: Vec<UninterpretedOption>,

}

Fields
java_package: Option<String>java_outer_classname: Option<String>java_multiple_files: Option<bool>java_generate_equals_and_hash: Option<bool>
๐Deprecated
java_string_check_utf8: Option<bool>optimize_for: Option<i32>go_package: Option<String>cc_generic_services: Option<bool>java_generic_services: Option<bool>py_generic_services: Option<bool>deprecated: Option<bool>cc_enable_arenas: Option<bool>objc_class_prefix: Option<String>csharp_namespace: Option<String>swift_prefix: Option<String>php_class_prefix: Option<String>php_namespace: Option<String>php_metadata_namespace: Option<String>ruby_package: Option<String>features: Option<FeatureSet>uninterpreted_option: Vec<UninterpretedOption>
Implementations

impl FileOptions

pub fn java_package(&self) -> &str

Returns the value of java_package, or the default value if java_package is unset.

pub fn java_outer_classname(&self) -> &str

Returns the value of java_outer_classname, or the default value if java_outer_classname is unset.

pub fn optimize_for(&self) -> OptimizeMode

Returns the enum value of optimize_for, or the default if the field is unset or set to an invalid enum value.

pub fn set_optimize_for(&mut self, value: OptimizeMode)

Sets optimize_for to the provided enum value.

pub fn java_multiple_files(&self) -> bool

Returns the value of java_multiple_files, or the default value if java_multiple_files is unset.

pub fn go_package(&self) -> &str

Returns the value of go_package, or the default value if go_package is unset.

pub fn cc_generic_services(&self) -> bool

Returns the value of cc_generic_services, or the default value if cc_generic_services is unset.

pub fn java_generic_services(&self) -> bool

Returns the value of java_generic_services, or the default value if java_generic_services is unset.

pub fn py_generic_services(&self) -> bool

Returns the value of py_generic_services, or the default value if py_generic_services is unset.

pub fn java_generate_equals_and_hash(&self) -> bool

Returns the value of java_generate_equals_and_hash, or the default value if java_generate_equals_and_hash is unset.

pub fn deprecated(&self) -> bool

Returns the value of deprecated, or the default value if deprecated is unset.

pub fn java_string_check_utf8(&self) -> bool

Returns the value of java_string_check_utf8, or the default value if java_string_check_utf8 is unset.

pub fn cc_enable_arenas(&self) -> bool

Returns the value of cc_enable_arenas, or the default value if cc_enable_arenas is unset.

pub fn objc_class_prefix(&self) -> &str

Returns the value of objc_class_prefix, or the default value if objc_class_prefix is unset.

pub fn csharp_namespace(&self) -> &str

Returns the value of csharp_namespace, or the default value if csharp_namespace is unset.

pub fn swift_prefix(&self) -> &str

Returns the value of swift_prefix, or the default value if swift_prefix is unset.

pub fn php_class_prefix(&self) -> &str

Returns the value of php_class_prefix, or the default value if php_class_prefix is unset.

pub fn php_namespace(&self) -> &str

Returns the value of php_namespace, or the default value if php_namespace is unset.

pub fn php_metadata_namespace(&self) -> &str

Returns the value of php_metadata_namespace, or the default value if php_metadata_namespace is unset.

pub fn ruby_package(&self) -> &str

Returns the value of ruby_package, or the default value if ruby_package is unset.