rumi_proto::xds::core

Module v3

Modules

collection_entry

Nested message and enum types in CollectionEntry.

resource_locator

Nested message and enum types in ResourceLocator.

Structs

Authority

xDS authority information.

CidrRange

CidrRange specifies an IP Address and a prefix length to construct the subnet mask for a CIDR <<https://tools.ietf.org/html/rfc4632>_> range.

CollectionEntry

xDS collection resource wrapper. This encapsulates a xDS resource when appearing inside a list collection resource. List collection resources are regular Resource messages of type:

ContextParams

Additional parameters that can be used to select resource variants. These include any global context parameters, per-resource type client feature capabilities and per-resource type functional attributes. All per-resource type attributes will be xds.resource. prefixed and some of these are documented below:

Resource

xDS resource wrapper. This encapsulates a xDS resource when appearing in an xDS transport discovery response or when accessed as a filesystem object.

ResourceLocator

xDS resource locators identify a xDS resource name and instruct the data-plane load balancer on how the resource may be located.

ResourceName

xDS resource name. This has a canonical xdstp:// URI representation:

TypedExtensionConfig

Message type for extension configuration.

Constants

FILE_DESCRIPTOR_SET

Encoded file descriptor set for the xds.core.v3 package