rumi_proto::xds::type::matcher

Module v3

Modules

double_range_matcher

Nested message and enum types in DoubleRangeMatcher.

int32_range_matcher

Nested message and enum types in Int32RangeMatcher.

int64_range_matcher

Nested message and enum types in Int64RangeMatcher.

ip_matcher

Nested message and enum types in IPMatcher.

matcher

Nested message and enum types in Matcher.

regex_matcher

Nested message and enum types in RegexMatcher.

server_name_matcher

Nested message and enum types in ServerNameMatcher.

string_matcher

Nested message and enum types in StringMatcher.

Structs

CelMatcher

Performs a match by evaluating a Common Expression Language <<https://github.com/google/cel-spec>_> (CEL) expression against the standardized set of :ref:HTTP attributes <arch_overview_attributes> specified via HttpAttributesCelMatchInput.

DoubleRangeMatcher

Specifies a set of ranges for matching a double number and the associated match actions.

HttpAttributesCelMatchInput

Specifies that matching should be performed on the set of :ref:HTTP attributes <arch_overview_attributes>.

Int32RangeMatcher

Specifies a set of ranges for matching an int32 number and the associated match actions.

Int64RangeMatcher

Specifies a set of ranges for matching an int64 number and the associated match actions.

IpMatcher

Matches a specific IP address against a set of possibly overlapping subnets using a trie.

ListStringMatcher

Specifies a list of ways to match a string.

Matcher

A matcher, which may traverse a matching tree in order to result in a match action. During matching, the tree will be traversed until a match is found, or if no match is found the action specified by the most specific on_no_match will be evaluated. As an on_no_match might result in another matching tree being evaluated, this process might repeat several times until the final OnMatch (or no match) is decided.

RegexMatcher

A regex matcher designed for safety when used with untrusted input.

ServerNameMatcher

Matches a fully qualified server name against a set of domain names with optional wildcards.

StringMatcher

Specifies the way to match a string. [#next-free-field: 9]

Constants

FILE_DESCRIPTOR_SET

Encoded file descriptor set for the xds.type.matcher.v3 package