rumi_proto::xds::type::matcher::v3::matcher

Module matcher_tree

Nested message and enum types in MatcherTree.

Structs

MatchMap

A map of configured matchers. Used to allow using a map within a oneof.

Enums

TreeType

Exact or prefix match maps in which to look up the input value. If the lookup succeeds, the match is considered successful, and the corresponding OnMatch is used.