rumi_proto::xds::type

Module v3

Modules

cel_expression

Nested message and enum types in CelExpression.

Structs

CelExpression

Either parsed or checked representation of the Common Expression Language <<https://github.com/google/cel-spec>_> (CEL) program.

CelExtractString

Extracts a string by evaluating a Common Expression Language <<https://github.com/google/cel-spec>_> (CEL) expression against the standardized set of :ref:HTTP attributes <arch_overview_attributes>.

DoubleRange

Specifies the double start and end of the range using half-open interval semantics [start, end).

Int32Range

Specifies the int32 start and end of the range using half-open interval semantics [start, end).

Int64Range

Specifies the int64 start and end of the range using half-open interval semantics [start, end).

TypedStruct

A TypedStruct contains an arbitrary JSON serialized protocol buffer message with a URL that describes the type of the serialized message. This is very similar to google.protobuf.Any, instead of having protocol buffer binary, this employs google.protobuf.Struct as value.

Constants

FILE_DESCRIPTOR_SET

Encoded file descriptor set for the xds.type.v3 package